* commit '0.3.4': (192 commits)
  0.3.4 release
  Fixed comment typo.
  Updated config documentation.
  Support for single 4D files in nipy.FitGLM
  BF: Identifying all 'mode options' of BET and make them mutually exclusive
  ENH: Complete BET interface. I'm not 100% sure if there was a reason to exclude the missing options, but they seems to work fine for me.
  BF: added package check for nipy interfaces
  BF: iterables order is always fixed by iterable and graph specification. it no longer depends on the random ordering of nodes at a given level.
  added commented test for iterable order need to turn into a real test
  Fixed smooth name formatting so smooth_iters works too
  Added dummy extension to split_filename doesn't strip measure
  BF: SurfXFM now handles surfaces with no file extension
  Added module imports for nipy.
  restored updatehash in run
  bf: config imported
  ENH: allowing relative paths as a configurable option
  ENH: debug option to create directories only
  allowing relative paths to be a configurable option
  fixed a bug with executable search and a local directory of the same name
  fixed fsl_resting tutorial
  ...

+++ nipype/interfaces/io.py
@@ -21,11 +21,13 @@ from copy import deepcopy
 import glob
 import os
 import shutil
+import hashlib
+import tempfile
 from warnings import warn
 
 from enthought.traits.trait_errors import TraitError
 try:
-    from xnatlib import Interface as XNATInterface
+    import pyxnat
 except:
     pass
 
@@ -41,6 +43,41 @@ from nipype.utils.filemanip import (copyfile, list_to_filename,
 
 import logging
 iflogger = logging.getLogger('interface')
+
+
+def copytree(src, dst):
+    """Recursively copy a directory tree using
+    nipype.utils.filemanip.copyfile()
+
+    This is not a thread-safe routine. However, in the case of creating new
+    directories, it checks to see if a particular directory has already been
+    created by another process.
+    """
+    names = os.listdir(src)
+    try:
+        os.makedirs(dst)
+    except OSError, why:
+        if 'File exists' in why:
+            pass
+        else:
+            raise why
+    errors = []
+    for name in names:
+        srcname = os.path.join(src, name)
+        dstname = os.path.join(dst, name)
+        try:
+            if os.path.isdir(srcname):
+                copytree(srcname, dstname)
+            else:
+                copyfile(srcname, dstname, True)
+        except (IOError, os.error), why:
+            errors.append((srcname, dstname, str(why)))
+        # catch the Error from the recursive copytree so that we can
+        # continue with other files
+        except Exception, err:
+            errors.extend(err.args[0])
+    if errors:
+        raise Exception, errors
 
 def add_traits(base, names, trait_type=None):
     """ Add traits to a traited class.
@@ -86,6 +123,8 @@ class DataSinkInputSpec(DynamicTraitedSpec):
                                          'to substitute and string to replace'
                                          'it with'))
     _outputs = traits.Dict(traits.Str, value={}, usedefault=True)
+    remove_dest_dir = traits.Bool(False, usedefault=True,
+                                  desc='remove dest directory when copying dirs')
     
     def __setattr__(self, key, value):
         if key not in self.copyable_trait_names():
@@ -108,7 +147,13 @@ class DataSink(IOBase):
 
         An attribute such as contrasts.@con will create a 'contrasts' directory to
         store the results linked to the attribute. If the @ is left out, such as in
-        'contrasts.con' a subdirectory 'con' will be created under 'contrasts'.  
+        'contrasts.con' a subdirectory 'con' will be created under 'contrasts'.
+
+        .. note::
+
+        Unlike most nipype-nodes this is not a thread-safe node because it can
+        write to a common shared location. It will not complain when it
+        overwrites a file.
 
         Examples
         --------
@@ -161,7 +206,13 @@ class DataSink(IOBase):
         if isdefined(self.inputs.container):
             outdir = os.path.join(outdir, self.inputs.container)
         if not os.path.exists(outdir):
-            os.makedirs(outdir)
+            try:
+                os.makedirs(outdir)
+            except OSError, inst:
+                if 'File exists' in inst:
+                    pass
+                else:
+                    raise(inst)
         for key,files in self.inputs._outputs.items():
             iflogger.debug("key: %s files: %s"%(key, str(files)))
             files = filename_to_list(files)
@@ -185,7 +236,13 @@ class DataSink(IOBase):
                     dst = self._substitute(dst)
                     path,_ = os.path.split(dst)
                     if not os.path.exists(path):
-                        os.makedirs(path)
+                        try:
+                            os.makedirs(path)
+                        except OSError, inst:
+                            if 'File exists' in inst:
+                                pass
+                            else:
+                                raise(inst)
                     iflogger.debug("copyfile: %s %s"%(src, dst))
                     copyfile(src, dst, copy=True)
                 elif os.path.isdir(src):
@@ -194,12 +251,18 @@ class DataSink(IOBase):
                     dst = self._substitute(dst)
                     path,_ = os.path.split(dst)
                     if not os.path.exists(path):
-                        os.makedirs(path)
-                    if os.path.exists(dst):
+                        try:
+                            os.makedirs(path)
+                        except OSError, inst:
+                            if 'File exists' in inst:
+                                pass
+                            else:
+                                raise(inst)
+                    if os.path.exists(dst) and self.inputs.remove_dest_dir:
                         iflogger.debug("removing: %s"%dst)
                         shutil.rmtree(dst)
                     iflogger.debug("copydir: %s %s"%(src, dst))
-                    shutil.copytree(src, dst)
+                    copytree(src, dst)
         return None
 
 
@@ -208,6 +271,8 @@ class DataGrabberInputSpec(DynamicTraitedSpec): #InterfaceInputSpec):
             desc='Path to the base directory consisting of subject data.')
     raise_on_empty = traits.Bool(True, usedefault=True,
                           desc='Generate exception if list is empty for a given field')
+    sort_filelist = traits.Bool(False, usedefault=True,
+                        desc='Sort the filelist that matches the template')
     template = traits.Str(mandatory=True,
              desc='Layout used to get files. relative to base directory if defined')
     template_args = traits.Dict(traits.Str,
@@ -343,6 +408,8 @@ class DataGrabber(IOBase):
                     else:
                         warn(msg)
                 else:
+                    if self.inputs.sort_filelist:
+                        filelist.sort()
                     outputs[key] = list_to_filename(filelist)
             for argnum, arglist in enumerate(args):
                 maxlen = 1
@@ -376,6 +443,8 @@ class DataGrabber(IOBase):
                             warn(msg)
                         outputs[key].insert(i, None)
                     else:
+                        if self.inputs.sort_filelist:
+                            outfiles.sort()
                         outputs[key].insert(i,list_to_filename(outfiles))
             if any([val==None for val in outputs[key]]):
                 outputs[key] = []
@@ -486,14 +555,21 @@ class FreeSurferSource(IOBase):
         
 
 class XNATSourceInputSpec(DynamicTraitedSpec): #InterfaceInputSpec):
-    config_file = File(exists=True, mandatory=True,
-                        desc='a json config file containing xnat access info: url, username and password')
     query_template = traits.Str(mandatory=True,
              desc='Layout used to get files. relative to base directory if defined')
     query_template_args = traits.Dict(traits.Str,
                                 traits.List(traits.List),
                                 value=dict(outfiles=[]), usedefault=True,
                                 desc='Information to plug into template')
+
+    xnat_server = traits.Str(mandatory=True, requires=['xnat_user', 'xnat_pwd'], xor=['xnat_config'])
+    xnat_user = traits.Str()
+    xnat_pwd = traits.Password()
+    xnat_config = File(mandatory=True, xor=['xnat_server'])
+
+    cache_dir = Directory(desc='Cache directory')
+    cache_size = traits.Str(desc='Optional cache max size')
+
 
 class XNATSource(IOBase):
     """ Generic XNATSource module that wraps around glob in an
@@ -576,9 +652,16 @@ class XNATSource(IOBase):
     def _list_outputs(self):
         # infields are mandatory, however I could not figure out how to set 'mandatory' flag dynamically
         # hence manual check
-        config_info = load_json(self.inputs.config_file)
-        cwd = os.getcwd()
-        xnat = XNATInterface(config_info['url'], config_info['username'], config_info['password'], cachedir=cwd)
+
+        cache_dir = self.inputs.cache_dir or tempfile.gettempdir()
+
+        if self.inputs.xnat_server:
+            xnat = pyxnat.Interface(self.inputs.xnat_server,self.inputs.xnat_user, self.inputs.xnat_pwd, cache_dir)
+        else:
+            xnat = pyxnat.Interface(self.inputs.xnat_config)
+
+#        xnat.set_offline_mode()
+
         if self._infields:
             for key in self._infields:
                 value = getattr(self.inputs,key)
@@ -596,7 +679,7 @@ class XNATSource(IOBase):
                     self.inputs.field_template.has_key(key):
                 template = self.inputs.field_template[key]
             if not args:
-                file_objects = xnat.select(template).request_objects()
+                file_objects = xnat.select(template).get('obj')
                 if file_objects == []:
                     raise IOError('Template %s returned no files'%template)
                 outputs[key] = list_to_filename([str(file_object.get()) for file_object in file_objects])
@@ -621,12 +704,12 @@ class XNATSource(IOBase):
                         else:
                             argtuple.append(arg)
                     if argtuple:
-                        file_objects = xnat.select(template%tuple(argtuple)).request_objects()
+                        file_objects = xnat.select(template%tuple(argtuple)).get('obj')
                         if file_objects == []:
                             raise IOError('Template %s returned no files'%(template%tuple(argtuple)))
                         outfiles = list_to_filename([str(file_object.get()) for file_object in file_objects])
                     else:
-                        file_objects = xnat.select(template).request_objects()
+                        file_objects = xnat.select(template).get('obj')
                         if file_objects == []:
                             raise IOError('Template %s returned no files'%template)
                         outfiles = list_to_filename([str(file_object.get()) for file_object in file_objects])
@@ -637,4 +720,100 @@ class XNATSource(IOBase):
                 outputs[key] = outputs[key][0]
         return outputs
 
+
+class XNATSinkInputSpec(DynamicTraitedSpec):
+    
+    _outputs = traits.Dict(traits.Str, value={}, usedefault=True)
+
+    xnat_server = traits.Str(mandatory=True, requires=['xnat_user', 'xnat_pwd'], xor=['xnat_config'])
+    xnat_user = traits.Str()
+    xnat_pwd = traits.Password()
+    xnat_config = File(mandatory=True, xor=['xnat_server'])
+    cache_dir = Directory(desc='')
+    cache_size = traits.Str()
+
+    project_id =  traits.Str(desc='Project in which to store the outputs', mandatory=True)
+    subject_id =  traits.Str(desc='Set to subject id', mandatory=True)
+    experiment_id =  traits.Str(desc='Set to workflow name', mandatory=True)
+
+    def __setattr__(self, key, value):
+        if key not in self.copyable_trait_names():
+            self._outputs[key] = value
+        else:
+            super(XNATSinkInputSpec, self).__setattr__(key, value)
+
+    
+class XNATSink(IOBase):
+    """ Generic datasink module that takes a directory containing a
+        list of nifti files and provides a set of structured output
+        fields.
+    """
+    input_spec = XNATSinkInputSpec
+
+    def _list_outputs(self):
+        """Execute this module.
+        """
+
+        cache_dir = self.inputs.cache_dir or tempfile.gettempdir()
+
+        if self.inputs.xnat_server:
+            xnat = pyxnat.Interface(self.inputs.xnat_server,self.inputs.xnat_user, self.inputs.xnat_pwd, cache_dir)
+        else:
+            xnat = pyxnat.Interface(self.inputs.xnat_config)
+
+        uri_template_args = {'project_id':self.inputs.project_id,
+                             'subject_id':'%s_%s' % \
+                                (self.inputs.project_id, 
+                                 self.inputs.subject_id),
+                             'experiment_id': '%s_%s' % \
+                                (hashlib.md5(self.inputs.subject_id).hexdigest(),
+                                 self.inputs.experiment_id)
+                             }
+
+        for key,files in self.inputs._outputs.items():
+            for name in filename_to_list(files):
+
+                if isinstance(name, list):
+                    for i, file_name in enumerate(name):
+                        write_file_to_XNAT(xnat, file_name, '%s_'%i+key, uri_template_args)
+
+                else:
+                    write_file_to_XNAT(xnat, name, key, uri_template_args)
+
+
+def write_file_to_XNAT(xnat, name, key, uri_template_args):
+
+    val_list = [val
+         for part in os.path.split(name)[0].split(os.sep) 
+         for val in part.split('_')[1:] 
+         if part.startswith('_') and len(part.split('_'))%2]
+
+    keymap = dict(zip(val_list[1::2],val_list[2::2]))
+
+    recon_label = []
+    for key, val in sorted(keymap.items()):
+        if str(self.inputs.subject_id) not in val:
+            recon_label.extend([key, val])
+
+    uri_template_args['recon_label'] = \
+        hashlib.md5(uri_template_args['experiment_id']).hexdigest()
+
+    if recon_label:
+        uri_template_args['recon_label'] += '_'.join(recon_label)
+
+    uri_template_args['resource_label'] = '%s_%s' % \
+                (hashlib.md5(uri_template_args['recon_label']).hexdigest(),
+                 key.split('.')[0]
+                 )
+
+    uri_template_args['file_name'] = os.path.split(os.path.abspath(name))[1]
+
+    uri_template = ('/project/%(project_id)s/subject/%(subject_id)s'
+                    '/experiment/%(experiment_id)s/reconstruction/%(recon_label)s'
+                    '/out/resource/%(resource_label)s/file/%(file_name)s')
+
+    print uri_template%uri_template_args
+
+    file_resource = xnat.select(uri_template%uri_template_args)
+    file_resource.put(name, experiments='xnat:imageSessionData')
